Move functions from dns package back into ACME.
